mostRecent() created

+++ HabitTracker/TodayView.swift
@@ -9,13 +9,18 @@
         - What habit have been logged & how many times
         - If 
  */
+import Foundation
 import SwiftUI
 import SwiftData
+import UIKit
 
 struct TodayView: View {
-    @State private var isShowingNewHabitForm = false
     @Environment(\.modelContext) private var context
+    
+    @State private var isShowingNewHabitForm = false
+    
     @Query private var habits: [Habit]
+    @Query private var loggedHabits: [LoggedHabit]
     
     var body: some View {
         ZStack{
@@ -30,21 +35,42 @@ struct TodayView: View {
                 Spacer()
                 // Most used habits
                 //displayMostUsed()
-                
-                List{
-                    ForEach(habits){ item in
-                        DisplayAllHabits(habit: item)
-                    }.onDelete { itemSet in
-                        for item in itemSet{
-                            context.delete(habits[item])
-                        }
+                if habits.isEmpty{
+                    Text("Add your first Habit")
+                }else{
+                    if loggedHabits.isEmpty{
+                        List{
+                            ForEach(habits){ item in
+                                DisplayAllHabits(habit: item)
+                            }.onDelete { itemSet in
+                                for item in itemSet{
+                                    context.delete(habits[item])
+                                }
+                            }
+                            
+                        }.preferredColorScheme(/*@START_MENU_TOKEN@*/.dark/*@END_MENU_TOKEN@*/)
+                            .scrollDisabled(true)
+                            .scrollContentBackground(.hidden)
+                            .listRowSpacing(10.0)
+                            .frame(height: UIScreen.main.bounds.height * 0.4)
+                    }else{
+                        let topFiveHabits: [Habit] = mostRecent(loggedHabits: loggedHabits, allHabits: habits)
+                        List{
+                            ForEach(topFiveHabits){ item in
+                                DisplayAllHabits(habit: item)
+                            }.onDelete { itemSet in
+                                for item in itemSet{
+                                    context.delete(habits[item])
+                                }
+                            }
+                            
+                        }.preferredColorScheme(/*@START_MENU_TOKEN@*/.dark/*@END_MENU_TOKEN@*/)
+                            .scrollDisabled(true)
+                            .scrollContentBackground(.hidden)
+                            .listRowSpacing(10.0)
+                            .frame(height: UIScreen.main.bounds.height * 0.4)
                     }
-                    
-                }.scrollDisabled(true)
-                    .scrollContentBackground(.hidden)
-                    .listRowSpacing(10.0)
-                    .frame(height: UIScreen.main.bounds.height * 0.4)
-                
+                }
                 
                 HStack(alignment: /*@START_MENU_TOKEN@*/.center/*@END_MENU_TOKEN@*/){
                     //button to create new habit
